+++ include/seqan3/alignment/aligned_sequence/debug_stream_alignment.hpp
@@ -18,11 +18,11 @@
 #include <range/v3/algorithm/for_each.hpp>
 
 #include <seqan3/alignment/aligned_sequence/aligned_sequence_concept.hpp>
-#include <seqan3/core/concept/tuple.hpp>
 #include <seqan3/core/detail/debug_stream_tuple.hpp>
 #include <seqan3/range/views/slice.hpp>
 #include <seqan3/range/views/to_char.hpp>
 #include <seqan3/range/views/zip.hpp>
+#include <seqan3/utility/tuple/concept.hpp>
 
 namespace seqan3::detail
 {
