+++ include/seqan3/alignment/pairwise/detail/concept.hpp
@@ -12,11 +12,11 @@
 
 #pragma once
 
+#include <seqan3/std/ranges>
 #include <tuple>
 
 #include <seqan3/alphabet/concept.hpp>
-#include <seqan3/core/concept/tuple.hpp>
-#include <seqan3/std/ranges>
+#include <seqan3/utility/tuple/concept.hpp>
 
 namespace seqan3::detail
 {
