+++ include/seqan3/core/detail/debug_stream_tuple.hpp
@@ -16,8 +16,8 @@
 #include <seqan3/std/ranges>
 
 #include <seqan3/alphabet/concept.hpp>
-#include <seqan3/core/concept/tuple.hpp>
 #include <seqan3/core/detail/debug_stream_type.hpp>
+#include <seqan3/utility/tuple/concept.hpp>
 
 namespace seqan3
 {
