+++ include/seqan3/io/detail/record.hpp
@@ -16,9 +16,9 @@
 
 #include <seqan3/std/ranges>
 
-#include <seqan3/core/concept/tuple.hpp>
 #include <seqan3/io/record.hpp>
 #include <seqan3/range/views/repeat.hpp>
+#include <seqan3/utility/tuple/concept.hpp>
 #include <seqan3/utility/type_list/traits.hpp>
 
 namespace seqan3::detail
