+++ include/seqan3/io/sequence_file/output.hpp
@@ -21,7 +21,6 @@
 #include <vector>
 
 #include <seqan3/core/detail/pack_algorithm.hpp>
-#include <seqan3/core/concept/tuple.hpp>
 #include <seqan3/io/stream/concept.hpp>
 #include <seqan3/io/exception.hpp>
 #include <seqan3/io/record.hpp>
@@ -39,6 +38,7 @@
 #include <seqan3/range/views/convert.hpp>
 #include <seqan3/range/views/get.hpp>
 #include <seqan3/range/views/zip.hpp>
+#include <seqan3/utility/tuple/concept.hpp>
 #include <seqan3/utility/type_list/traits.hpp>
 
 namespace seqan3
