+++ test/unit/alignment/pairwise/align_pairwise_test.cpp
@@ -7,6 +7,7 @@
 
 #include <gtest/gtest.h>
 
+#include <seqan3/std/ranges>
 #include <stdexcept>
 #include <type_traits>
 #include <utility>
@@ -16,10 +17,9 @@
 #include <seqan3/alignment/pairwise/align_pairwise.hpp>
 #include <seqan3/alphabet/gap/gapped.hpp>
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
-#include <seqan3/core/concept/tuple.hpp>
 #include <seqan3/range/views/to_char.hpp>
 #include <seqan3/range/views/to.hpp>
-#include <seqan3/std/ranges>
+#include <seqan3/utility/tuple/concept.hpp>
 
 using seqan3::operator""_dna4;
