+++ test/unit/alphabet/composite/alphabet_tuple_base_test_template.hpp
@@ -8,7 +8,7 @@
 #include <gtest/gtest.h>
 
 #include <seqan3/alphabet/composite/alphabet_tuple_base.hpp>
-#include <seqan3/core/concept/tuple.hpp>
+#include <seqan3/utility/tuple/concept.hpp>
 
 using seqan3::get;
