+++ test/unit/io/record_test.cpp
@@ -12,11 +12,11 @@
 
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/alphabet/quality/phred42.hpp>
-#include <seqan3/core/concept/tuple.hpp>
 #include <seqan3/core/detail/debug_stream_alphabet.hpp>
 #include <seqan3/io/detail/record.hpp>
 #include <seqan3/io/record.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
+#include <seqan3/utility/tuple/concept.hpp>
 
 using seqan3::operator""_dna4;
